requirements.txt
numpy
pandas
matplotlib
datasets
pytorch
scikit-learn
sentence_transformers
tqdm
scipy